+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "example-crates/keychain_tracker_example_cli",
     "example-crates/wallet_electrum",
     "example-crates/wallet_esplora",
+    "example-crates/wallet_esplora_async",
     "nursery/tmp_plan",
     "nursery/coin_select"
 ]
