+++ example-crates/wallet_esplora_async/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "bdk-esplora-wallet-example"
+name = "wallet_esplora_async"
 version = "0.1.0"
 edition = "2021"
-publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 bdk = { path = "../../crates/bdk" }
-bdk_esplora = { path = "../../crates/esplora", features = ["blocking"] }
+bdk_esplora = { path = "../../crates/esplora", features = ["async-https"] }
 bdk_file_store = { path = "../../crates/file_store" }
+tokio = { version = "1", features = ["rt", "rt-multi-thread", "macros"] }
